requirements.txt
openai
pydub
tqdm
python-dotenv
requests
